requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.9
aiohttp==3.11.10
Pillow==11.0.0
python-dotenv==1.0.0
opencv-python>=4.8.0
numpy>=1.26.0
torch>=2.0.0
# SAM2 - install separately with:
# pip install git+https://github.com/facebookresearch/segment-anything-2.git

# Handwriting Recognition & Font Matching:
# TrOCR (Microsoft Transformer-based OCR for handwriting):
# Vision Transformer (ViT) for advanced font matching:
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
# Note: Already have Pillow and opencv-python above

# Alternative OCR options (not needed if using TrOCR):
# pip install paddlepaddle paddleocr
# pip install easyocr
